Added two calls to obj->reset_init_m_if_stateless()
When XML parsing input stream tags "read" and "object" or "read", "objects" and "object", call the obj->reset_init_m_if_stateless() function. This is to check if obj->init_m flag should be cleared in order to trigger a read from the bus instead of returning a buffered value.

The decision depends on the object flag "s" for stateless in the linknx.conf. If the "s" object flag is not set, linknx will delivere a buffered value to the user without causing any bus activity. On the other hand, if "s" object flag is set, a bus read is triggered every time the value gets queried from linknx.