configure.ac: always define CPPUNIT
If the user gives --without-cppunit to the configure script, build fails
on:
configure: error: conditional "CPPUNIT" was never defined.
Usually this means the macro was only invoked conditionally.

This is due to the fact that CPPUNIT is only defined if cppunit is set
to auto. To fix this error, also define CPPUNIT if the user provides
--with-cppunit or --without-cppunit option

Fixes:
 - http://autobuild.buildroot.org/results/4b042b7e18425690ec26b4977865516bedcb9edb

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>